requirements.txt
streamlit
PyQt5
pywebview